Moved specs into subdirectories
To match the names in the source directory.
