+++ specs/contentTypes/ContentDataTypeChecksSpec.ts
@@ -1,5 +1,5 @@
-import { ContentDataTypes } from "../src/contentTypes/ContentDataTypes";
-import { ContentDataTypeChecks } from "../src/contentTypes/ContentDataTypeChecks";
+import { ContentDataTypes } from "../../src/contentTypes/ContentDataTypes";
+import { ContentDataTypeChecks } from "../../src/contentTypes/ContentDataTypeChecks";
 
 describe("ContentDataTypeChecks", function () {
   it("returns true when included is undefined and excluded is undefined", async function () {
