+++ specs/tilesetData/TilesetSourceSpec.ts
@@ -1,9 +1,9 @@
-import { TilesetSource } from "../src/tilesetData/TilesetSource";
-import { TilesetSourceFs } from "../src/tilesetData/TilesetSourceFs";
-import { TilesetInMemory } from "../src/tilesetData/TilesetInMemory";
+import { TilesetSource } from "../../src/tilesetData/TilesetSource";
+import { TilesetSourceFs } from "../../src/tilesetData/TilesetSourceFs";
+import { TilesetInMemory } from "../../src/tilesetData/TilesetInMemory";
 
-import { TilesetSource3tz } from "../src/packages/TilesetSource3tz";
-import { TilesetSource3dtiles } from "../src/packages/TilesetSource3dtiles";
+import { TilesetSource3tz } from "../../src/packages/TilesetSource3tz";
+import { TilesetSource3dtiles } from "../../src/packages/TilesetSource3dtiles";
 
 async function createTilesetInMemory() {
   const tileset = new TilesetInMemory();
